requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
python-multipart==0.0.9
aiofiles==23.2.1